+++ MelonLoader/AssertionManager.h
@@ -12,6 +12,7 @@ public:
 
 	static void Start(const char* filename, const char* position);
 	static void ThrowError(std::string msg, const char* filepath = NULL);
+	static void ThrowInternalError(std::string msg);
 	static void Decide(void* thing, const char* name);
 	static HMODULE LoadLib(const char* name, const char* filepath);
 	static HMODULE GetModuleHandlePtr(const char* name);