+++ MelonLoader/Exports.cpp
@@ -4,6 +4,7 @@
 #include "Mono.h"
 #include "HookManager.h"
 #include "Logger.h"
+#include "AssertionManager.h"
 
 void Log(MonoString* namesection, MonoString* txt) { Logger::Log(Mono::mono_string_to_utf8(namesection), Mono::mono_string_to_utf8(txt)); }
 void LogColor(MonoString* namesection, MonoString* txt, ConsoleColor color) { Logger::Log(Mono::mono_string_to_utf8(namesection), Mono::mono_string_to_utf8(txt), color); }
@@ -31,6 +32,7 @@ MelonLoader::LoadMode GetLoadMode_Mods() { return MelonLoader::LoadMode_Mods; }
 bool AG_Force_Regenerate() { return MelonLoader::AG_Force_Regenerate; }
 MonoString* AG_Force_Version_Unhollower() { if (MelonLoader::ForceUnhollowerVersion != NULL) return Mono::mono_string_new(Mono::Domain, MelonLoader::ForceUnhollowerVersion); return NULL; }
 void SetTitleForConsole(MonoString* txt) { Console::SetTitle(Mono::mono_string_to_utf8(txt)); }
+void ThrowInternalError(MonoString* txt) { AssertionManager::ThrowInternalError(Mono::mono_string_to_utf8(txt)); }
 
 void Exports::AddInternalCalls()
 {
@@ -64,6 +66,7 @@ void Exports::AddInternalCalls()
 	Mono::mono_add_internal_call("MelonLoader.MelonLogger::Native_LogError", LogError);
 	Mono::mono_add_internal_call("MelonLoader.MelonLogger::Native_LogMelonError", LogMelonError);
 	Mono::mono_add_internal_call("MelonLoader.MelonLogger::Native_LogMelonCompatibility", LogMelonCompatibility);
+	Mono::mono_add_internal_call("MelonLoader.MelonLogger::Native_ThrowInternalError", ThrowInternalError);
 
 	Mono::mono_add_internal_call("MelonLoader.AssemblyGenerator::Force_Regenerate", AG_Force_Regenerate);
 	Mono::mono_add_internal_call("MelonLoader.AssemblyGenerator::Force_Version_Unhollower", AG_Force_Version_Unhollower);